Reverted DOTApp and Fee source modifications

+++ ethereum/contracts/FeeSource.sol
@@ -4,5 +4,5 @@ pragma experimental ABIEncoderV2;
 
 // Something that can burn a fee from a feepayer account.
 interface FeeSource {
-    function burnFee(address feePayer, uint128 _amount) external;
+    function burnFee(address feePayer, uint256 _amount) external;
 }
