+++ ethereum/contracts/test/MockFeeSource.sol
@@ -5,7 +5,7 @@ pragma experimental ABIEncoderV2;
 import "../FeeSource.sol";
 
 contract MockFeeSource is FeeSource {
-    function burnFee(address, uint128 _amount) pure external override {
+    function burnFee(address, uint256 _amount) pure external override {
         // Simulate the case where the user has no funds,
         require(_amount != 1024, "User has no funds to burn");
     }
